+++ CHANGELOG.md
@@ -8,6 +8,7 @@ OJP-Demo URL: https://opentdatach.github.io/ojp-demo-app/
 - Fix station board stage changes - [PR #372](https://github.com/openTdataCH/ojp-demo-app-src/pull/372)
   - fix station board change stage bug
   - when changing stages do a BBOX LIR to fetch the stop nearby from the new stage
+  - keep current stage model in one place (UserTripService) - [PR #373](https://github.com/openTdataCH/ojp-demo-app-src/pull/373)
 
 20.March 2026
 - Improve TripInfo state - [PR #371](https://github.com/openTdataCH/ojp-demo-app-src/pull/371)